# RD-9528: Monkeypatched Kiama files + improvements to transpilation time