--- a/.scalafmt.conf
+++ b/.scalafmt.conf
@@ -1,6 +1,8 @@
 version = 2.7.5
 maxColumn = 120
 
+project.excludeFilters = ["/inkytonik/.*"]
+
 optIn.forceBlankLineBeforeDocstring = true
 trailingCommas = "never"