--- a/raw-compiler-rql2-truffle/src/test/scala/raw/compiler/rql2/tests/regressions/TruffleRegressionTest.scala
+++ b/raw-compiler-rql2-truffle/src/test/scala/raw/compiler/rql2/tests/regressions/TruffleRegressionTest.scala
@@ -100,5 +100,4 @@ import raw.testing.tags.TruffleTests
 @TruffleTests class RD9229TruffleTest extends TruffleCompilerTestContext with RD9229Test
 @TruffleTests class RD9409TruffleTest extends TruffleCompilerTestContext with RD9409Test
 @TruffleTests class RD9479TruffleTest extends TruffleCompilerTestContext with RD9479Test
-@TruffleTests class RD9528TruffleTest extends TruffleCompilerTestContext with RD9528Test
 @TruffleTests class RD9554TruffleTest extends TruffleCompilerTestContext with RD9554Test