… raw-compiler-rql2/src/test/scala/raw/compiler/rql2/tests/regressions/RD9528Test.scala