@miguelbranco80 2023-09-13
So we go from 6.7s to 0.9s per query 👍